Cropped location file to only include areas affected by hurricanes, thus eliminating majority of South America.
Here now long/lat coordinates are named. A challenge still exists where some localities are being recorded outside land borders so we are unable to apply a country name in some cases and this is a problem as these are populated centroids. A gBuffer type additions might help the problem because these locations are not a far way off land.

An attempt is bing made with a shape file of Puerto Rico to collect hurricane impact by year.